added modal and Carousel to weekly component
