+++ src/components/pages/Gallery/Weekly.js
@@ -1,8 +1,25 @@
 import React, { useState } from 'react';
 import { connect } from 'react-redux';
-import CarouselContainer from './CarouselContainer';
+import { Modal, Carousel } from 'antd';
 
 const Weekly = props => {
+  const [isModalVisible, setIsModalVisible] = useState(false);
+  const [imgUrl, setImgUrl] = useState([]);
+
+  const showModal = link => {
+    setIsModalVisible(true);
+    setImgUrl(link);
+    console.log('clicked: ', link);
+  };
+
+  const handleCancel = () => {
+    setIsModalVisible(false);
+  };
+
+  function onChange(a, b, c) {
+    console.log(a, b, c);
+  }
+
   return (
     <>
       <div className="weekly-sub-container">
@@ -16,6 +33,7 @@ const Weekly = props => {
               className="gallery-submission"
               src={props.drawing}
               alt="drawing submision"
+              onClick={() => showModal([props.drawing])}
             />
           </div>
           <div className="sub-container">
@@ -23,10 +41,29 @@ const Weekly = props => {
               className="gallery-submission"
               src={props.writing}
               alt="writing submision"
+              onClick={() => showModal([props.writing])}
             />
           </div>
         </span>
       </div>
+      <Modal
+        visible={isModalVisible}
+        centered
+        onCancel={handleCancel}
+        footer={null}
+      >
+        <Carousel afterChange={onChange} arrows={true}>
+          {imgUrl.map(url => (
+            <div>
+              <img
+                style={{ height: '72vh', objectFit: 'contain' }}
+                alt=""
+                src={url}
+              />
+            </div>
+          ))}
+        </Carousel>
+      </Modal>
     </>
   );
 };
